refactor(ir): actually remove `analysis.py`
